feat: update home cw components
