+++ apps/mobile/src/components/home/ContinueWatchingList.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 
 import { IContinueWatching } from '@naikamu/shared';
-import { useNavigation } from '@react-navigation/native';
-import { FlatList, Pressable, StyleSheet, Text, View } from 'react-native';
-import { ProgressBar } from 'react-native-paper';
+import { FlatList, StyleSheet, Text, View } from 'react-native';
 
 import { useQueryGetContinueWatching } from '../../api/hooks';
 import { useTranslate } from '../../i18n/useTranslate';
-import { RootStackScreenNames, SeriesStackScreenNames } from '../../routes';
-import { colors, defaultRadius, fontStyles } from '../../styles';
-import { ProgressiveImage } from '../ProgressiveImage';
+import { colors, fontStyles, globalStyle } from '../../styles';
+
+import { HomeEpisodeElement } from './HomeEpisodeElement';
 
 export type ContinueWatchingListProps = Record<string, never>;
 
@@ -18,11 +16,13 @@ export const ContinueWatchingList: React.FC<
 > = ({}) => {
   const { data, refetch, isRefetching } = useQueryGetContinueWatching();
   const { translate } = useTranslate();
-  const { navigate } = useNavigation<any>();
 
   return (
     <View>
-      <Text>{translate('continue watching')}</Text>
+      <Text
+        style={[fontStyles.header, colors.textLight, globalStyle.marginBottom]}>
+        {translate('home.headers.continueWatching')}
+      </Text>
       {data ? (
         <FlatList
           data={data}
@@ -30,38 +30,7 @@ export const ContinueWatchingList: React.FC<
           onRefresh={refetch}
           refreshing={isRefetching}
           renderItem={({ item }: { item: IContinueWatching }) => (
-            <Pressable
-              onPress={() =>
-                navigate(RootStackScreenNames.SeriesStack, {
-                  screen: SeriesStackScreenNames.Episodes,
-                  params: {
-                    seriesId: item.anime.id,
-                  },
-                })
-              }
-              style={styles.mainContainer}>
-              <View>
-                <View style={[styles.posterContainer]}>
-                  <ProgressiveImage
-                    resizeMode="cover"
-                    source={item.anime.poster}
-                    style={styles.poster}
-                  />
-                </View>
-                <ProgressBar
-                  progress={item.episode.progress / (24 * 60)}
-                  style={styles.progressBar}
-                  theme={{
-                    colors: {
-                      primary: colors.accent.color,
-                    },
-                  }}
-                />
-              </View>
-              <Text style={[fontStyles.normal, colors.textLight]}>
-                Episode: {item.episode.number}
-              </Text>
-            </Pressable>
+            <HomeEpisodeElement item={item} />
           )}
         />
       ) : null}
@@ -69,24 +38,4 @@ export const ContinueWatchingList: React.FC<
   );
 };
 
-const styles = StyleSheet.create({
-  mainContainer: {
-    width: 100,
-    height: 200,
-    flexDirection: 'column',
-    marginRight: 20,
-  },
-  posterContainer: {
-    width: '100%',
-    height: 160,
-  },
-  poster: {
-    width: '100%',
-    height: '100%',
-    borderTopLeftRadius: defaultRadius,
-    borderTopRightRadius: defaultRadius,
-  },
-  progressBar: {
-    zIndex: 1,
-  },
-});
+const styles = StyleSheet.create({});
